% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationsfunctions.R
\name{pois_dist_fct}
\alias{pois_dist_fct}
\title{Probability mass function for a Poisson distribution shifted by one and resulting expected value 8}
\usage{
pois_dist_fct(x)
}
\arguments{
\item{x}{vector of positive integer values to evaluate}
}
\description{
Probability mass function for a Poisson distribution shifted by one and resulting expected value 8
}
\examples{

plot(pois_dist_fct(1:100))

}
