% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevtoinc.R
\docType{package}
\name{prevtoinc}
\alias{prevtoinc}
\alias{prevtoinc-package}
\title{prevtoinc:}
\description{
The \strong{prevtoinc} package implements functionality to simulate 
point-prevalence studies (PPSs) for healtcare-associated infections (HAIs) and estimate incidence based on the PPS data. It 
is a companion package to the preprint Willrich et al., From prevalence to incidence - a new approach in the hospital setting;
<doi:10.1101/554725> , where methods are explained in detail.
}
